Ver1.1.4  （UI及功能优化  增加代金券功能）

See merge request win/winpos!10

+++ WinSaasPOS/Model/QueryOrder.cs
@@ -79,7 +79,7 @@ namespace WinSaasPOS.Model
 
          public string title{get;set;}
 
-         public decimal num{get;set;}
+         public int num{get;set;}
 
          public decimal specnum{get;set;}
